+++ requirements.txt
@@ -1,44 +1,8 @@
-aniso8601==8.0.0
-appdirs==1.4.4
-attrs==20.2.0
-black==20.8b1
-certifi==2020.6.20
-chardet==3.0.4
-click==7.1.1
-dataclasses==0.7
-flake8==3.8.4
 Flask==1.1.2
-Flask-Cors==3.0.9
-Flask-RESTful==0.3.8
-Flask-SQLAlchemy==2.4.1
-idna==2.10
-importlib-metadata==2.0.0
-iniconfig==1.1.1
-itsdangerous==1.1.0
-Jinja2==2.11.1
-MarkupSafe==1.1.1
-mccabe==0.6.1
-mypy-extensions==0.4.3
-packaging==20.4
-pathspec==0.8.0
-pluggy==0.13.1
-psycopg2-binary==2.8.6
-py==1.9.0
-pycodestyle==2.6.0
-pyflakes==2.2.0
-pyparsing==2.4.7
-pytest==6.1.1
-python-decouple==3.3
-python-dotenv==0.14.0
-pytz==2020.1
-regex==2020.10.15
-requests==2.24.0
-six==1.15.0
-SQLAlchemy==1.3.16
-toml==0.10.1
-typed-ast==1.4.1
-typing-extensions==3.7.4.3
-urllib3==1.25.11
-Werkzeug==1.0.1
-zipp==3.3.1
+requests~=2.24.0
+setuptools~=50.3.2
+waitress
 coverage
+flask_sqlalchemy
+flask_cors
+psycopg2
